Fixed #7 added tests and fixed #27 bug on X-Content-Type plugin.
In this commit, a set of functional tests have been written for
X-Content-Type plugin.

Furthermore, we have fixed bug #27 in XContentTypePlugin. We now
consider both upper- and lower-case header name. At this point, we are
case-insensitive when it comes to checking the value of the header. This
is related to ticket #28 and shall be discussed.

Lastly, we also added a function to kill ports (1234, 1235, 1443) in
``setUpClass`` and ``tearDownClass``. Apparently, some of the test
classes were unable to be killed properly from calling ``.terminate()``.
This will be revisited at some point.