Working on #36 adding test and validator to robots plugin.
Currently robots plugin does not actually verify the
existence of robots.txt. If the target url is 200-reacable, plugin
will return success. This is address by ticket #36.

Ticket #11 addresses the lack of validator in plugin. We should verify
the robots.txt is valid and is confronting to the standard.

We probably will adaopt the recommendations from the official
Robots.org, in addition to w3c.

[1]: http://www.robotstxt.org/orig.html

[2]: http://www.w3.org/TR/html4/appendix/notes.html#h-B.4.1.1