Fixed #50, #51, #52, #56. Added sites to groups collections.
Problem:

Most of these issues is caused by sites are not being added groups when
a site is being created. Some part of the documentation was not clear
either. I also overlooked at some of the issues.

Solution:

This commit appends the site url to each group
specified by ``groups``. Furthermore, the ``_call(..)`` function in the
function test has been modified to remove 'application/json' from GET
request to prevent Flask from complaining.

Problems remain to be solved:

1. When a group is created, owner should be included automatically. This
is trivial fix; require user to be added as part of the request, and
add user to the default list instead of an ``[]`` list.

2. Review documentations.

3. We still want a few more individual functional tests for
/report/issues.

4. We still have to fix #55.