Fixed #109. Users collection has a status field.
We now have an option to either mark a user active via direct signup, or
mark a user as invited via invitation. Bad users can be marked as
'banned'. When a user is marked as invited, the user is given a url
(front-end) to login. Upon the first login, the status will change to
'active'. Ticket #111 will check user's status and prevent banned users
(and non-active) users from accessing Minion.