Fixed #140. Fixed #142. See #146. Added basic permission check to scans.
In order to allow the backend to verify permission, a user query string is appended. This query string is '?email=<email-address>'. The check is performed by finding {'user': email, 'site': url} in the groups collections. Currently, issue #147 is reported but given our model, #147 is not a security issue. Therefore, by knowing a user (his/her Persona account) belongs to at least one group that contains the target url is sufficent enough to say the user has permission to view the scan or do work on the scan.