Allow to specify a whitelist and a blacklist of net access. See #217.
This is my last pull request as an intern on the Mozilla Security
Assurance team.

What hasn't been implemented is the abililty to resolve the target
to the IP addres within the scan.

This is due to time constraint; I need more time
to develop a test case (maybe I overthink, maybe just localhost
is a fine test case, maybe). It is relatively very easy to add that
check anyway. I already have the resolver function implemented
in utils.py

The changes are to allow a "scna.json". We will extend the network
access before saving the target url to database in the future
commits.

The test includes two sample JSON. One with both blacklist and whitelist
empted and one with a blacklist. We are using ipaddress to determine
whether or not an ip is part of a whitelist (remind: anything
not in blacklist is automatically consider whitelisted).

I also want to be able to specify ports in the future (whitelist
a particular port on 192.168.1.1)).

We can do this once this implemented is accepted.

Anyway. I have had a great time at Mozilla. See you on #websectools.

*rawr*