Working on #284. Update test_hsts.py to use new driver.
+ Replace stunnel with Flask SSL invokcation.

This snippet (http://flask.pocoo.org/snippets/111/)
can be used to start Flask with TLS. A new directory called https_data
has been added to replace stunnel-data. The new direcorty contains
a testing certificate and a testing private key. There is a REAMDE
with instruction to generate the cert and key.

Unfortunately, because we self-signed the cert, we need to disable SSL/TLS
cert verification in pycurl. But I don't want to enable that until I have
spoken to Stefan about this. For now, we know the code works because the plugin
hasn't changed a bit. We will leave it commented out. I really hope we don't
need to assert mozillalabs.com or google.com to test the plugin.