Fixed #295. Fixed running plugin tests as group.
The regression happens because in the old revision each plugin test
had a different endpoint. Since now they (almost all of the tests)
still share the same Flask object and use the same endpoint ``/tests``,
when nosetests scans for all the test_app object, only one of the endpoint
is chosen. Sometimes if we are lucky on tearDown some are updated. But
in general, only one test file's actual endpoint is in the memory.

The solution is to pull all the Flask-related code into a separate python
file according to test. They now organized under servers/ directory.